add docker-registry PR-961 as a patch file
The PR was merged after the final version was released, and therefore
is not part of any docker registry release for us to use.

We hit the same issues that this PR fixes on our registry causing
multiple errors on our logs and finally causing our registry to freeze.

Patch removes a ParallelKey class that is used with big uploads to s3 to
increase performance. The PR had found that it causes problems, and at
the same time doesn't actually improve performance on the tests he did.