Update registry to 2.6.2
This update includes a security fix for CVE-2017-11468. Release notes from github/docker/distribution for this version:

This release is a special security release to address an issue allowing an attacker to force arbitrarily-sized memory allocations in a registry instance through the manifest endpoint. The problem has been mitigated by limiting the size of reads for image manifest content.

Change-Type: minor
Connects-To: #26